requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
easydict==1.13
eval_type_backport==0.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.20.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
prettytable==3.12.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
simsimd==6.0.6
six==1.16.0
stringzilla==3.10.10
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
